Convert to plugin
